% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool-package.R
\docType{package}
\name{slidetool-package}
\alias{slidetool-package}
\title{slidetool: Retrieve Data from Whole Slide Images}
\description{
Uses OpenSlide to retrieve data from whole slide images, also known as virtual slide files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/slidetool}
  \item Report bugs at \url{https://github.com/luciorq/slidetool/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [copyright holder]

}
\keyword{internal}
