% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool.R
\name{slidetool}
\alias{slidetool}
\title{Run OpenSlide Command-Line Tool}
\usage{
slidetool(...)
}
\arguments{
\item{...}{Additional arguments passed to the \code{slidetool} command-line tool.}
}
\value{
A list containing the result of the \code{slidetool} command execution, including standard output and error messages.
}
\description{
Executes the OpenSlide command-line tool \code{slidetool} with the specified arguments.
This function uses the \code{condathis} package to manage the execution environment.
}
\details{
The \code{slidetool()} function serves as a wrapper for the OpenSlide \code{slidetool} command.
It allows you to execute any \code{slidetool} subcommands by passing them as arguments.
}
\examples{
\dontrun{
# List associated images in a slide
slidetool("assoc", "list", "path/to/slide.svs")
}

}
