% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool_assoc.R
\name{slidetool_assoc_icc_read}
\alias{slidetool_assoc_icc_read}
\title{Read ICC Color Profile of an Associated Image}
\usage{
slidetool_assoc_icc_read(path, assoc)
}
\arguments{
\item{path}{Character string specifying the path to the slide file.}

\item{assoc}{Character string specifying the name of the associated image.}
}
\value{
A list containing the result of the \code{slidetool} command execution.
}
\description{
Writes the ICC color profile of a specified associated image to an output file or standard output.
If no ICC color profile is available, the function will produce an error.
}
\details{
This function retrieves the ICC color profile of an associated image within a slide using OpenSlide's \code{slidetool}.
}
\examples{
\dontrun{
# Read ICC profile of an associated image
icc_profile <- slidetool_assoc_icc_read("path/to/slide.svs", "label")
}

}
