% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool_assoc.R
\name{slidetool_assoc_list}
\alias{slidetool_assoc_list}
\title{List Associated Images in a Slide}
\usage{
slidetool_assoc_list(path)
}
\arguments{
\item{path}{Character string specifying the path to the slide file.}
}
\value{
A character vector containing the names of associated images.
}
\description{
Retrieves a list of associated images for a slide. Associated images are small non-pyramidal images containing additional information, such as labels or thumbnails.
}
\details{
This function extracts the names of associated images embedded within a whole slide image file using OpenSlide.
}
\examples{
\dontrun{
# List associated images
associated_images <- slidetool_assoc_list("path/to/slide.svs")
print(associated_images)
}

}
