% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool_assoc.R
\name{slidetool_assoc_read}
\alias{slidetool_assoc_read}
\title{Read an Associated Image from a Slide}
\usage{
slidetool_assoc_read(path, assoc, output_file = NULL)
}
\arguments{
\item{path}{Character string specifying the path to the slide file.}

\item{assoc}{Character string specifying the name of the associated image.}

\item{output_file}{Character string specifying the path to the output PNG file. If \code{NULL}, the image will be written to a temporary file output.}
}
\description{
Extracts a specific associated image from a slide and saves it as a PNG image.
}
\examples{
\dontrun{
# Read ICC profile of an associated image
thumbnail_path <- slidetool_assoc_read("path/to/slide.svs", "thumbnail")
}

}
