% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool_prop.R
\name{slidetool_prop_get}
\alias{slidetool_prop_get}
\title{Get a Slide Property Value}
\usage{
slidetool_prop_get(path, property)
}
\arguments{
\item{path}{Character string specifying the path to the slide file.}

\item{property}{Character string specifying the name of the property to retrieve.}
}
\value{
A character string containing the value of the specified property.
}
\description{
Retrieves the value of a specific OpenSlide property for a given slide.
}
\details{
This function allows you to access individual pieces of metadata by specifying the property name.
}
\examples{
\dontrun{
# Get a specific property value
vendor <- slidetool_prop_get("path/to/slide.svs", "openslide.vendor")
print(vendor)
}

}
