% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidetool_prop.R
\name{slidetool_prop_list_names}
\alias{slidetool_prop_list_names}
\title{List Slide Property Names}
\usage{
slidetool_prop_list_names(path)
}
\arguments{
\item{path}{Character string specifying the path to the slide file.}
}
\value{
A character vector containing the names of all properties.
}
\description{
Retrieves the names of all OpenSlide properties for a given slide.
}
\details{
This function provides the names of all available properties without their values, useful for exploring available metadata.
}
\examples{
\dontrun{
# Get property names
property_names <- slidetool_prop_list_names("path/to/slide.svs")
print(property_names)
}

}
